Refactor src/db.rs: Extract persistence logic to modules
- Move PersistenceTracker to src/storage/index_persistence/tracker.rs
- Move spawn_background_persistence_thread to src/storage/index_persistence/worker.rs
- Move persist_* and load_* functions to src/storage/index_persistence/operations.rs
- Clean up src/db.rs to reduce God Object smell
- Ensure API compatibility and pass all tests
