+++ .jules/forge.md
@@ -1,0 +1,3 @@
+## 2024-05-23 - src/db.rs is a God Object
+**Learning:** `src/db.rs` contains over 1900 lines and mixes core database API logic with persistence management details (Tracker, Background Thread, Helper functions). This violates Single Responsibility Principle.
+**Action:** Extract persistence logic into `src/storage/index_persistence/` modules (`tracker`, `worker`, `operations`).
